pyproject.toml
[project]
name = "teamtalk-reg-system"
version = "0.2.0"
description = "Telegram bot and FastAPI web app for TeamTalk server self-registration."
authors = [
    {name = "Kirill", email = "kirillkolovi@gmail.com"},
]
requires-python = ">=3.11"
license = {file = "LICENSE"}
readme = "README.md"

# Основные зависимости проекта
dependencies = [
    "aiogram ~=3.20.0",
    "py-talk-ex ~=1.6.0",
    "python-dotenv ~=1.1.0",
    "SQLAlchemy ~=2.0.41",
    "aiosqlite ~=0.21.0",
    "fastapi ~=0.115",
    "uvicorn[standard] ~=0.34.3",
    "Babel ~=2.17.0",
    "Jinja2 ~=3.1.6",
    "python-multipart ~=0.0.20",
    "aiofiles==24.1.0",
]

[project.urls]
Homepage = "https://github.com/kirill-jjj/teamtalk_reg_system"
Repository = "https://github.com/kirill-jjj/teamtalk_reg_system"
"Bug Tracker" = "https://github.com/kirill-jjj/teamtalk_reg_system/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
backend-path = "."
[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]
exclude = ["locales*"]
